feat(dpav-1441): moved forms input container
